web_radius/requirements.txt
asgiref==3.7.2
autopep8==1.6.0
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==0.5.0
Django==4.1
django-environ==0.8.1
djangorestframework==3.14.0
drf-api-logger==1.1.14
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.6
pycodestyle==2.8.0
python-decouple==3.8
pytz==2021.3
requests==2.31.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2023.3
Unipath==1.1
urllib3==2.0.3
whitenoise==5.3.0
